@@ -1,20 +1,21 @@
 {
-  "name": "antd-demo-ts",
-  "version": "0.1.0",
+  "name": "sandra.frontend",
+  "version": "0.3.0",
   "private": true,
   "dependencies": {
-    "antd": "^3.3.1",
     "axios": "^0.18.0",
     "color-hash": "^1.0.3",
-    "javascript-time-ago": "^1.0.28",
-    "mobx": "^4.1.0",
-    "mobx-react": "^5.0.0",
-    "mobx-react-router": "^4.0.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "javascript-time-ago": "^1.0.30",
+    "mobx": "^4.3.0",
+    "mobx-react": "^5.1.2",
+    "mobx-react-router": "^4.0.4",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
-    "react-scripts-ts": "2.13.0",
+    "react-scripts-ts": "2.16.0",
+    "semantic-ui-css": "2.2.12",
+    "semantic-ui-react": "^0.80.2",
     "store": "^2.0.12",
     "timeago.js": "^3.0.2"
   },
@@ -25,12 +26,12 @@
     "eject": "react-scripts-ts eject"
   },
   "devDependencies": {
-    "@types/jest": "^22.1.3",
-    "@types/node": "^9.4.6",
-    "@types/react": "^16.0.40",
-    "@types/react-dom": "^16.0.4",
-    "@types/react-router": "^4.0.23",
-    "@types/react-router-dom": "^4.2.5",
-    "typescript": "^2.7.2"
+    "@types/jest": "^22.2.3",
+    "@types/node": "^10.1.2",
+    "@types/react": "^16.3.14",
+    "@types/react-dom": "^16.0.5",
+    "@types/react-router": "^4.0.25",
+    "@types/react-router-dom": "^4.2.6",
+    "typescript": "^2.8.3"
   }
 }
