@@ -1,4 +1,4 @@
-export { default as Message } from './Message';
 export { default as UserCard } from './User';
+export { default as MessageSimple } from './MessageSimple';
+export { default as MessageComplex } from './MessageComplex';
 export { default as Subscription } from './Subscription';
-export { default as AddSubscriptionDialog } from './AddSubscriptionDialog';
