@@ -1,123 +1,139 @@
 import * as React from 'react';
 import { inject, observer } from 'mobx-react';
 
-import { Form, Input, Icon, Button, message } from 'antd';
-import { FormComponentProps } from 'antd/lib/form/Form';
-import { Link } from 'react-router-dom';
+import {
+  Button, Form, Message, Segment,
+  InputOnChangeData
+} from 'semantic-ui-react';
 
 import ApiClient from '../lib/client';
 import { RouterStore } from 'mobx-react-router';
+import SessionState from '../lib/SessionStore';
+import { User } from '../lib/api_interfaces';
 
-const FORM_FIELDS = {
-  EMAIL: 'email',
-  PASSWORD: 'password',
-  PASSWORD_AGAIN: 'password2'
-};
-
-interface Props extends FormComponentProps {
+interface Props {
   routing?: RouterStore;
+  session?: SessionState;
 }
 interface States {
+  nickname: string;
+  email: string;
+  password: string;
+  password_confirm: string;
   loading: boolean;
+  error: boolean;
+  message: string;
+  user?: User;
 }
 
-@inject('routing')
+@inject('routing', 'session')
 @observer
 class RegForm extends React.Component<Props, States> {
-  state = {
-    loading: false
+  state: States = {
+    nickname: '',
+    email: '',
+    password: '',
+    password_confirm: '',
+    loading: false,
+    error: false,
+    message: ''
   };
 
-  handleSubmit: React.FormEventHandler<void> = (e) => {
-    const { push } = this.props.routing!;
-    this.setState({ loading: true });
+  handleChange = (_: React.SyntheticEvent<HTMLInputElement>, data: InputOnChangeData) =>
+    this.setState({ [data.name]: data.value })
+  handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    const { email, password, password_confirm, nickname } = this.state;
+    this.setState({ loading: true, error: false });
     e.preventDefault();
-    this.props.form.validateFields((err, values) => {
-      if (!err) {
-        ApiClient.register(values[FORM_FIELDS.EMAIL], values[FORM_FIELDS.PASSWORD])
-          .then((user) => {
-            message.info(<p>注册成功，用户 id {user.id}。</p>);
-            push('/login');
-          })
-          .catch((ex) => {
-            message.error(<p>{ex.message} - {ex.response && ex.response.data && ex.response.data.code}</p>);
-          })
-          .then(() => this.setState({ loading: false }));
+    try {
+      if (password !== password_confirm) {
+        throw new Error('两次输入的密码不一致');
       }
-    });
-  }
-  // tslint:disable-next-line:no-any
-  compareToFirstPassword = (rule: any, value: any, callback: any) => {
-    const form = this.props.form;
-    if (value && value !== form.getFieldValue(FORM_FIELDS.PASSWORD)) {
-      callback('两次输入的密码不一致');
-    } else {
-      callback();
+      const user = await ApiClient.register(email, password, nickname);
+      this.setState({ user });
+      setTimeout(() => this.props.routing!.push('/login'), 1000);
+    } catch (ex) {
+      this.setState({
+        error: true,
+        message: (ex.response && ex.response.data && ex.response.data.code) || ex.message
+      });
     }
+    this.setState({ loading: false });
   }
   render() {
-    const { getFieldDecorator, getFieldsError, isFieldTouched } = this.props.form;
-    const fieldsError = getFieldsError();
+    const {
+      nickname,
+      email,
+      password,
+      password_confirm,
+      loading,
+      error,
+      message,
+      user
+    } = this.state;
     return (
-      <Form onSubmit={this.handleSubmit} >
-        <h2>注册</h2>
-        <Form.Item>
-          {getFieldDecorator(FORM_FIELDS.EMAIL, {
-            rules: [
-              { type: 'email', message: '请输入有效的邮箱地址' },
-              { required: true, message: '请输入邮箱' }
-            ],
-          })(
-            <Input
-              type="email"
-              prefix={<Icon type="mail" style={{ color: 'rgba(0,0,0,.25)' }} />}
-              placeholder="邮箱"
-            />
-          )}
-        </Form.Item>
-        <Form.Item>
-          {getFieldDecorator(FORM_FIELDS.PASSWORD, {
-            rules: [{ required: true, message: '请输入密码' }],
-          })(
-            <Input
-              prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />}
-              type="password"
-              placeholder="密码"
-            />
-          )}
-        </Form.Item>
-        <Form.Item>
-          {getFieldDecorator(FORM_FIELDS.PASSWORD_AGAIN, {
-            rules: [
-              { required: true, message: '请再次输入密码' },
-              { validator: this.compareToFirstPassword }
-            ],
-          })(
-            <Input
-              prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />}
-              type="password"
-              placeholder="确认密码"
-            />
-          )}
-        </Form.Item>
-        <Form.Item>
-          <Button
-            type="primary"
-            htmlType="submit"
-            disabled={
-              Object.keys(fieldsError).some(field => fieldsError[field]) ||
-              !(isFieldTouched(FORM_FIELDS.EMAIL) && isFieldTouched(FORM_FIELDS.PASSWORD))
-            }
-            loading={this.state.loading}
-          >
-            注册
-          </Button>
-        </Form.Item>
-        <Link to="/login" >已有账号，登入</Link>
+      <Form
+        size="large"
+        onSubmit={this.handleSubmit}
+        loading={loading}
+        error={error}
+        success={user !== undefined}
+      >
+        <Segment raised>
+          <Message success>
+            <Message.Header>注册成功</Message.Header>
+            <p>您的用户代码：{user && user.id}，3 秒后将为您导向登入页面。</p>
+          </Message>
+          <Message
+            error
+            header="注册失败"
+            content={message}
+          />
+          <Form.Input
+            fluid
+            icon="user"
+            iconPosition="left"
+            placeholder="昵称"
+            type="text"
+            name="nickname"
+            value={nickname}
+            onChange={this.handleChange}
+          />
+          <Form.Input
+            fluid
+            icon="mail"
+            iconPosition="left"
+            placeholder="邮箱"
+            type="email"
+            name="email"
+            value={email}
+            onChange={this.handleChange}
+          />
+          <Form.Input
+            fluid
+            icon="lock"
+            iconPosition="left"
+            placeholder="密码"
+            type="password"
+            name="password"
+            value={password}
+            onChange={this.handleChange}
+          />
+          <Form.Input
+            fluid
+            icon="lock"
+            iconPosition="left"
+            placeholder="确认密码"
+            type="password"
+            name="password_confirm"
+            value={password_confirm}
+            onChange={this.handleChange}
+          />
+          <Button primary fluid size="large">注册</Button>
+        </Segment>
       </Form>
     );
   }
 }
-const DecoratedRegForm = Form.create()(RegForm);
 
-export default DecoratedRegForm;
+export default RegForm;
