@@ -1,3 +1,4 @@
 export { default as Login } from './Login';
 export { default as Reg } from './Reg';
-export { default as AddSubscription } from './AddSubscription';
+export { default as CreateSubscription } from './CreateSubscription';
+export { default as UpdateSubscription } from './UpdateSubscription';
