@@ -1,14 +1,12 @@
 import * as React from 'react';
 import { inject, observer } from 'mobx-react';
-import { RouterStore, syncHistoryWithStore } from 'mobx-react-router';
-import { Router, Route, Switch, Redirect, Link } from 'react-router-dom';
-import SessionState from '../lib/state/Session';
-import * as Interfaces from '../lib/api_interfaces';
+import { RouterStore } from 'mobx-react-router';
+import { Route, Switch } from 'react-router-dom';
+import SessionState from '../lib/SessionStore';
 
-import * as Forms from '../Forms';
 import * as Components from '../Components';
 import * as DashboardComponents from './DashboardPages';
-import { Row, Col, Button, Collapse, Menu, Icon, message } from 'antd';
+import { Grid, Menu, MenuItemProps, Header, Image, Segment } from 'semantic-ui-react';
 
 interface Props {
   routing?: RouterStore;
@@ -23,47 +21,45 @@ class Dashboard extends React.Component<Props, States> {
   async componentWillMount() {
     try {
       await this.props.session!.loadSession();
-      await this.props.session!.retrieveLatestData();
     } catch (ex) {
-      message.error(<p>{ex.message} - {ex.response && ex.response.data && ex.response.data.code}</p>);
+      // message.error(<p>{ex.message} - {ex.response && ex.response.data && ex.response.data.code}</p>);
     }
   }
+  handleItemClick: MenuItemProps['onClick'] = (_, { route }) => this.props.routing!.push(route || '/dashboard');
   render() {
-    const { messages, subscriptions } = this.props.session!;
     return (
-      <Row style={{ height: '100%' }} type="flex" justify="space-around" align="top">
-        <Col xl={18} lg={22} xs={24}>
-          <h1 style={{ marginTop: '16px', marginBottom: '16px' }}>Sandra</h1>
-          <Row>
-            <Col md={15} xs={24}>
-              <Menu
-                mode="horizontal"
-                selectedKeys={[this.props.routing!.location!.pathname]}
-              >
-                <Menu.Item key="/dashboard">
-                  <Link to="/dashboard">首页</Link>
-                </Menu.Item>
-                <Menu.Item key="/dashboard/subscriptions">
-                  <Link to="/dashboard/subscriptions">订阅</Link>
-                </Menu.Item>
-              </Menu>
-            </Col>
-          </Row>
-          <Row>
-            <Col md={15} xs={24}>
-              <Switch>
-                <Route path="/dashboard" exact={true} component={DashboardComponents.Messages} />
-                <Route path="/dashboard/subscriptions" component={DashboardComponents.Subscriptions} />
-              </Switch>
-            </Col>
-            <Col md={9} xs={24}>
-              <Row>
-                <Components.UserCard />
-              </Row>
-            </Col>
-          </Row>
-        </Col>
-      </Row>
+      <Grid columns={2} container>
+        <Grid.Row verticalAlign="top">
+          <Grid.Column width={3}>
+            <Segment>
+              <Header as="h1">
+                <Image src="/assets/oin.svg" /> beta
+              </Header>
+            </Segment>
+            <Components.UserCard />
+            <Menu pointing secondary vertical fluid>
+              <Menu.Item
+                name="dashboard"
+                route="/dashboard"
+                active={this.props.routing!.location!.pathname === '/dashboard'}
+                onClick={this.handleItemClick}
+              />
+              <Menu.Item
+                name="subscriptions"
+                route="/dashboard/subscriptions"
+                active={this.props.routing!.location!.pathname === '/dashboard/subscriptions'}
+                onClick={this.handleItemClick}
+              />
+            </Menu>
+          </Grid.Column>
+          <Grid.Column width={13}>
+            <Switch>
+              <Route path="/dashboard" exact={true} component={DashboardComponents.Messages} />
+              <Route path="/dashboard/subscriptions" component={DashboardComponents.Subscriptions} />
+            </Switch>
+          </Grid.Column>
+        </Grid.Row>
+      </Grid>
     );
   }
 }
