@@ -1,11 +1,12 @@
 import * as React from 'react';
-import { Row, Col } from 'antd';
 
-import * as Forms from '../Forms';
+import { Grid, Header, Message } from 'semantic-ui-react';
+import { Link } from 'react-router-dom';
+
 import EnsureAnonymous from '../Routes/EnsureAnonymous';
+import * as Forms from '../Forms';
 
 interface Props {
-
 }
 interface States {
 }
@@ -13,12 +14,21 @@ interface States {
 class RegPage extends React.Component<Props, States> {
   render() {
     return (
-      <Row style={{ height: '100%' }} type="flex" justify="space-around" align="middle">
-        <Col md={12} xs={24}>
-          <EnsureAnonymous />
+      <Grid
+        textAlign="center"
+        style={{ height: '100%' }}
+        verticalAlign="middle"
+      >
+        <Grid.Column style={{ maxWidth: 450 }}>
+          <Header as="h2" textAlign="center">
+            注册<EnsureAnonymous />
+          </Header>
           <Forms.Reg />
-        </Col>
-      </Row>
+          <Message>
+            已有账户？<Link to="/login">登入</Link>
+          </Message>
+        </Grid.Column>
+      </Grid>
     );
   }
 }
