@@ -1,8 +1,8 @@
 import * as React from 'react';
-import { Route, Redirect } from 'react-router-dom';
+import { Redirect } from 'react-router-dom';
 import { inject, observer } from 'mobx-react';
 
-import SessionState from '../lib/state/Session';
+import SessionState from '../lib/SessionStore';
 
 @inject('session')
 @observer
@@ -11,7 +11,7 @@ export default class EnsureAnonymous extends React.Component<{
   // tslint:disable-next-line:no-any
 } & any> {
   render() {
-    const { session, ...restProps } = this.props;
+    const { session } = this.props;
     if (session!.authenticated) { return <Redirect to="/dashboard" />; }
     return <div />;
   }
