@@ -2,8 +2,8 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
-import 'antd/dist/antd.css';
 import './image.css';
+import 'semantic-ui-css/semantic.min.css';
 
 ReactDOM.render(
   <App />,
